requirements.txt
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
diffusers==0.26.3
transformers==4.38.2
accelerate==0.28.0
huggingface_hub==0.21.4
peft==0.10.0
torch==2.2.0
tqdm==4.66.1